Separate things into bin/lib crates
